Move things defined in `atl` but actually used in `transform`. This reduces the coupling of two modules, but not eleminates it, as `transform` still references `ATLTransformBase` at module init time.
